Adding converters
Former-commit-id: 99db7cb3fc44d5f60fd4f73c0a796c7291ee5e4b

+++ fortscale/presidio-input/presidio-input-core/src/main/java/presidio/input/core/RawEventsPageIterator.java
@@ -30,7 +30,7 @@ public class RawEventsPageIterator<U extends AbstractPresidioDocument> implement
         this.pageSize = pageSize;
         this.startDate = startDate;
         this.endDate = endDate;
-        long totalNumberOfEvents = presidioInputPersistencyService.count(schema, startDate, endDate);
+        float totalNumberOfEvents = presidioInputPersistencyService.count(schema, startDate, endDate);
         this.totalAmountOfPages = (int) Math.ceil(totalNumberOfEvents / pageSize);
     }
